Create a separate taxonomy validator in the infrastructure layer
